update tests
move tests to ignite.handlers along with fixtures and changing the
imports

add a test to ensure deprecation warnings for ignite.contrib.handlers
